+++ tools/airbyte_ci_code_validator/airbyte_ci_sonar_qube/__init__.py
@@ -4,7 +4,7 @@
 import os
 from pathlib import Path
 
-from ci_common_utils import Logger
+from airbyte_ci_common_utils import Logger
 
 LOGGER = Logger()
