+++ tools/airbyte_ci_code_validator/airbyte_ci_sonar_qube/main.py
