+++ tools/airbyte_ci_code_validator/setup.py
@@ -7,7 +7,7 @@ from setuptools import find_packages, setup
 
 MAIN_REQUIREMENTS = [
     "requests",
-    "ci_common_utils",
+    "airbyte_ci_common_utils",
     "unidiff",
     "mdutils~=1.3.1",
     "mypy==0.930",
@@ -17,7 +17,7 @@ TEST_REQUIREMENTS = ["requests-mock", "pytest", "black", "lxml", "isort"]
 
 setup(
     version="0.0.0",
-    name="ci_code_validator",
+    name="airbyte_ci_code_validator",
     description="Load and extract CI secrets for test suites",
     author="Airbyte",
     author_email="contact@airbyte.io",
@@ -29,8 +29,8 @@ setup(
     },
     entry_points={
         "console_scripts": [
-            "ci_sonar_qube = ci_sonar_qube.main:main",
-            "ci_changes_detection = ci_changes_detection.main:main",
+            "airbyte_ci_sonar_qube = airbyte_ci_sonar_qube.main:main",
+            "airbyte_ci_changes_detection = airbyte_ci_changes_detection.main:main",
         ],
     },
 )
