+++ tools/airbyte_ci_code_validator/tests/simple_files/mypy_smell_package_report.json
