+++ tools/airbyte_ci_code_validator/tests/simple_files/without_issues_report.json
