+++ tools/airbyte_ci_code_validator/tests/simple_smell_package/invalid_file.py
