+++ tools/airbyte_ci_credentials/README.md
@@ -10,7 +10,7 @@ Set up the world the same way Google Actions does it in `test-command.yml`.
 source venv/bin/activate
 pip install --quiet tox==3.24.4
 tox -r -c ./tools/tox_ci.ini
-pip install --quiet -e ./tools/ci_*
+pip install --quiet -e ./tools/airbyte_ci_*
 ```
 
 Download a Service account json key that has access to Google Secrets Manager.
@@ -19,7 +19,7 @@ Download a Service account json key that has access to Google Secrets Manager.
 export GCP_GSM_CREDENTIALS=`cat ~/Downloads/key.json`
 ```
 
-After making a change, you have to reinstall it to run the bash command: `pip install --quiet -e ./tools/ci_*`
+After making a change, you have to reinstall it to run the bash command: `pip install --quiet -e ./tools/airbyte_ci_*`
 
 ## Run it
 
@@ -28,11 +28,11 @@ The `VERSION=dev` will make it so it knows to use your local current working dir
 Pass in a connector name. For example:
 
 ```bash
-VERSION=dev ci_credentials destination-snowflake
+VERSION=dev airbyte_ci_credentials destination-snowflake
 ```
 
 To make sure it get's all changes every time, you can run this:
 
 ```bash
-pip install --quiet -e ./tools/ci_* && VERSION=dev ci_credentials destination-snowflake
+pip install --quiet -e ./tools/airbyte_ci_* && VERSION=dev airbyte_ci_credentials destination-snowflake
 ```
